requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cloudpickle==2.0.0
cmake==3.21.2
decorator==5.1.0
distro==1.6.0
dm-tree==0.1.6
flatbuffers==2.0
gast==0.4.0
google-auth==2.0.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.39.0
gviz-api==1.9.0
h5py==3.1.0
idna==3.2
jsonslicer==0.1.7
keras-nightly==2.8.0.dev2021100807
Keras-Preprocessing==1.1.2
libclang==11.1.0
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-build==0.12.0
six==1.15.0
tb-nightly==2.7.0a20210904
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.5.0
tensorboard-plugin-wit==1.8.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021102008
tf-nightly==2.8.0.dev20211020
tfa-nightly==0.15.0.dev20211014164307
tfp-nightly==0.15.0.dev20211008
tqdm==4.62.2
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zstandard==0.15.2
